Added turn directions #145
Fixes #167
Messed with #138 for the overloads for turn direction
