+++ src/EZ-Template/drive/drive.cpp
@@ -387,87 +387,3 @@ void Drive::initialize() {
   drive_imu_calibrate();
   drive_sensor_reset();
 }
-
-void Drive::pid_drive_toggle(bool toggle) { drive_toggle = toggle; }
-void Drive::pid_print_toggle(bool toggle) { print_toggle = toggle; }
-
-bool Drive::pid_drive_toggle_get() { return drive_toggle; }
-bool Drive::pid_print_toggle_get() { return print_toggle; }
-
-void Drive::slew_drive_constants_forward_set(okapi::QLength distance, int min_speed) {
-  double dist = distance.convert(okapi::inch);
-  slew_forward.constants_set(dist, min_speed);
-}
-
-void Drive::slew_drive_constants_backward_set(okapi::QLength distance, int min_speed) {
-  double dist = distance.convert(okapi::inch);
-  slew_backward.constants_set(dist, min_speed);
-}
-
-void Drive::slew_drive_constants_set(okapi::QLength distance, int min_speed) {
-  slew_drive_constants_backward_set(distance, min_speed);
-  slew_drive_constants_forward_set(distance, min_speed);
-}
-
-void Drive::slew_turn_constants_set(okapi::QAngle distance, int min_speed) {
-  double dist = distance.convert(okapi::degree);
-  slew_turn.constants_set(dist, min_speed);
-}
-
-void Drive::slew_swing_constants_backward_set(okapi::QLength distance, int min_speed) {
-  slew_swing_rev_using_angle = false;
-  double dist = distance.convert(okapi::inch);
-  slew_swing_backward.constants_set(dist, min_speed);
-}
-
-void Drive::slew_swing_constants_forward_set(okapi::QLength distance, int min_speed) {
-  slew_swing_fwd_using_angle = false;
-  double dist = distance.convert(okapi::inch);
-  slew_swing_forward.constants_set(dist, min_speed);
-}
-
-void Drive::slew_swing_constants_set(okapi::QLength distance, int min_speed) {
-  slew_swing_constants_forward_set(distance, min_speed);
-  slew_swing_constants_backward_set(distance, min_speed);
-}
-
-void Drive::slew_swing_constants_backward_set(okapi::QAngle distance, int min_speed) {
-  slew_swing_rev_using_angle = true;
-  double dist = distance.convert(okapi::degree);
-  slew_swing_backward.constants_set(dist, min_speed);
-}
-
-void Drive::slew_swing_constants_forward_set(okapi::QAngle distance, int min_speed) {
-  slew_swing_fwd_using_angle = true;
-  double dist = distance.convert(okapi::degree);
-  slew_swing_forward.constants_set(dist, min_speed);
-}
-
-void Drive::slew_swing_constants_set(okapi::QAngle distance, int min_speed) {
-  slew_swing_constants_forward_set(distance, min_speed);
-  slew_swing_constants_backward_set(distance, min_speed);
-}
-
-// Global enables for drive slew
-void Drive::slew_drive_set(bool slew_on) {
-  global_forward_drive_slew_enabled = slew_on;
-  global_backward_drive_slew_enabled = slew_on;
-}
-void Drive::slew_drive_forward_set(bool slew_on) { global_forward_drive_slew_enabled = slew_on; }
-bool Drive::slew_drive_forward_get() { return global_forward_drive_slew_enabled; }
-void Drive::slew_drive_backward_set(bool slew_on) { global_backward_drive_slew_enabled = slew_on; }
-bool Drive::slew_drive_backward_get() { return global_backward_drive_slew_enabled; }
-
-// Global enables for swing slew
-void Drive::slew_swing_set(bool slew_on) {
-  global_forward_swing_slew_enabled = slew_on;
-  global_backward_swing_slew_enabled = slew_on;
-}
-void Drive::slew_swing_forward_set(bool slew_on) { global_forward_swing_slew_enabled = slew_on; }
-bool Drive::slew_swing_forward_get() { return global_forward_swing_slew_enabled; }
-void Drive::slew_swing_backward_set(bool slew_on) { global_backward_swing_slew_enabled = slew_on; }
-bool Drive::slew_swing_backward_get() { return global_backward_swing_slew_enabled; }
-
-// Global enables for turn slew
-void Drive::slew_turn_set(bool slew_on) { global_turn_slew_enabled = slew_on; }
-bool Drive::slew_turn_get() { return global_turn_slew_enabled; }
